[#1144] Allow theming
This allows changing a few links and images in os-admin. It's not possible to
change the colorschema yet because I think the way to do this is to overwrite
SCSS variables, but I'm not sure exactly how.

As this is just the initial code to get the whitelabeling feature moving
forward, I think it's OK not to allow changing colors for now, and develop it
when we're implementing a new theme.

Fixes openspending/openspending#1144