Add /config endpoint to express server
This serves the configuration needed for the Angular app, adding
settings values from the environment or .env file as appropriate.

We use lookenv (https://github.com/RodrigoEspinosa/lookenv) to define
which env vars should be added, (and validate env vars which are
required). These are defined in ./lookenv.config.js.

Part of openspending/openspending#1340.