Rewrite origin url to remove /final/
The datapackage in /final/ isn't reflective of the originally uploaded
datapackage, and paths to resources are relative to
`data/pathtoresource.csv` which isn't compatible with the edit process.
this commit removes `final` from the url, so the package to be edited is
the one on the top level of the package directory in the datastore.